+++ packages/twenty-front/src/modules/object-record/record-field/ui/meta-types/display/components/RelationFromManyFieldDisplay.tsx
@@ -10,6 +10,7 @@ import { isActivityTargetField } from '@/object-record/record-field-list/utils/c
 import { FieldContext } from '@/object-record/record-field/ui/contexts/FieldContext';
 import { useFieldFocus } from '@/object-record/record-field/ui/hooks/useFieldFocus';
 import { useRelationFromManyFieldDisplay } from '@/object-record/record-field/ui/meta-types/hooks/useRelationFromManyFieldDisplay';
+import { ForbiddenFieldDisplay } from '@/object-record/record-field/ui/meta-types/display/components/ForbiddenFieldDisplay';
 import { extractTargetRecordsFromJunction } from '@/object-record/record-field/ui/utils/junction/extractTargetRecordsFromJunction';
 import { getJunctionConfig } from '@/object-record/record-field/ui/utils/junction/getJunctionConfig';
 import { hasJunctionConfig } from '@/object-record/record-field/ui/utils/junction/hasJunctionConfig';
@@ -143,6 +144,10 @@ export const RelationFromManyFieldDisplay = () => {
         return { record: extracted.record, objectMetadata };
       })
       .filter(isDefined);
+
+    if (fieldValue.some(isDefined) && targetRecordsWithMetadata.length === 0) {
+      return <ForbiddenFieldDisplay />;
+    }
 
     return (
       <ExpandableList isChipCountDisplayed={isFocused}>
