+++ packages/twenty-front/src/modules/object-record/record-table/record-table-cell/components/RecordTableCellFieldContextGeneric.tsx
@@ -6,13 +6,15 @@ import { type RecordField } from '@/object-record/record-field/types/RecordField
 import { FieldContext } from '@/object-record/record-field/ui/contexts/FieldContext';
 import { isFieldRelationManyToOne } from '@/object-record/record-field/ui/types/guards/isFieldRelationManyToOne';
 import { isFieldRelationOneToMany } from '@/object-record/record-field/ui/types/guards/isFieldRelationOneToMany';
+import { getJunctionConfig } from '@/object-record/record-field/ui/utils/junction/getJunctionConfig';
+import { getTargetObjectMetadataIdsFromField } from '@/object-record/record-field/ui/utils/junction/getTargetObjectMetadataIdsFromField';
 import { hasJunctionConfig } from '@/object-record/record-field/ui/utils/junction/hasJunctionConfig';
-import { isJunctionTargetForbidden } from '@/object-record/record-field/ui/utils/junction/isJunctionTargetForbidden';
 import { useRecordIndexContextOrThrow } from '@/object-record/record-index/contexts/RecordIndexContext';
 import { RecordUpdateContext } from '@/object-record/record-table/contexts/EntityUpdateMutationHookContext';
 import { useRecordTableContextOrThrow } from '@/object-record/record-table/contexts/RecordTableContext';
 import { useRecordTableRowContextOrThrow } from '@/object-record/record-table/contexts/RecordTableRowContext';
 import { useContext, type ReactNode } from 'react';
+import { isDefined } from 'twenty-shared/utils';
 
 type RecordTableCellFieldContextGenericProps = {
   recordField: RecordField;
@@ -61,13 +63,28 @@ export const RecordTableCellFieldContextGeneric = ({
       hasObjectReadPermissions &&
       hasJunctionConfig(fieldDefinition.metadata.settings)
     ) {
-      hasObjectReadPermissions = !isJunctionTargetForbidden({
+      const junctionConfig = getJunctionConfig({
         settings: fieldDefinition.metadata.settings,
-        junctionObjectMetadataId: relationObjectMetadataId,
+        relationObjectMetadataId,
         sourceObjectMetadataId: objectMetadataItem.id,
         objectMetadataItems,
-        objectPermissionsByObjectMetadataId,
       });
+
+      if (isDefined(junctionConfig)) {
+        const targetObjectMetadataIds = junctionConfig.targetFields.flatMap(
+          getTargetObjectMetadataIdsFromField,
+        );
+
+        if (targetObjectMetadataIds.length > 0) {
+          hasObjectReadPermissions = targetObjectMetadataIds.some(
+            (targetId) =>
+              getObjectPermissionsForObject(
+                objectPermissionsByObjectMetadataId,
+                targetId,
+              ).canReadObjectRecords,
+          );
+        }
+      }
     }
   }
